rename package to term
