+++ term/go110.go
@@ -1,6 +1,6 @@
 // +build go1.10
 
-package osc
+package term
 
 import (
 	"os"
